+++ packages/visual-editing/src/enable-visual-editing.ts
@@ -1,0 +1,10 @@
+export type {
+  DisableVisualEditing,
+  HistoryAdapter,
+  HistoryAdapterNavigate,
+  HistoryRefresh,
+  HistoryUpdate,
+  SuspiciousStegaReport,
+  VisualEditingOptions,
+} from './types'
+export {enableVisualEditing} from './ui/enableVisualEditing'